Self-defense against pump issues
We've had reports of partitions wedging and making no forward progress.
This only seems to happen to people who have topic claim mode active,
but partition mode health checking may just be masking it.

This patch does two things:

* Watch the last time we saw a message. The health check loop will now
release partition claims if the consumer velocity hits 0 and we are not
caught up. This is true even in topic claim mode.

* This also fixes a (potential) issue where updateOffsets is failing
on calling PartitionOffsets (a Kafka call). In that mode, we'd never
updateCurrentOffsets and the heartbeat loop would continue happily. If
this was happening I'd have expected log entries, but let's make this
more failure-proof anyway by checking errors.