Fix marshal data race
The data race is triggered because some partition claims are released in an asynchronous way. This allows
the possibility where the offending channel is closed while there is a claim still trying to write to it.
Here we add a wait group for every asynchronous claim release, so as to ensure all claims are stopped
before we close the channel.